Remove none working iframe workaround.
This hack caused vimb staying in input mode if the focus was changed by
clicking out of the form field. Now we have back issue with lost
inputmode in case the focused form field is within an iframe. But this
is the less common case and can by worked around by clicking into the
field twice.